// utils/api/types/education.types.ts
export type Education = {
  degree: string;
  institute: string;
  location: string;
  startYear: string;
  endYear: string;
  grade: string;
};

export type WorkExperience = {
  company: string;
  role: string;
  duration: string;
  description: string;
  logo?: string;
};

export type Project = {
  title: string;
  description: string;
};

export type ResumeData = {
  title?: string;
  fullName: string;
  location?: string;
  email?: string;
  phone?: string;
  objective?: string;
  portfolioLink?: string;
  githubLink?: string;
  linkedinLink?: string;
  technicalSkills?: string[];
  softSkills?: string[];
  educations?: Education[];
  experiences?: WorkExperience[];
  projects?: Project[];
};